requirements.txt
Django==5.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
django-cors-headers==4.3.1
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
